requirements-pyinstaller.txt
altgraph==0.17.4
asttokens==2.4.1
certifi==2025.10.5
charset-normalizer==3.4.3
cheap_repr==0.5.2
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.0
coverage==7.10.7
cramjam==2.11.0
cycler==0.12.1
deepmerge==1.1.1
dill==0.4.0
exceptiongroup==1.3.0
executing==2.2.1
fastparquet==2024.11.0
flake8==7.3.0
fonttools==4.60.1
frosch==0.1.9
fsspec==2025.9.0
future==1.0.0
greenlet==3.2.4
hyperopt==0.2.7
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
joblib==1.2.0
kiwisolver==1.4.7
kneed==0.8.5
lightgbm==3.3.5
matplotlib==3.9.4
mccabe==0.7.0
multiprocess==0.70.18
networkx==3.2.1
numpy==1.26.4
numpy-indexed==0.3.7
packaging==25.0
pandas==2.3.3
pefile==2023.2.7
pillow==11.3.0
pluggy==1.6.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.0
psycopg2==2.9.10
pure_eval==0.2.3
py==1.11.0
py4j==0.10.9.9
pyarrow==15.0.2
pycodestyle==2.14.0
pydocstyle==6.3.0
pyflakes==3.4.0
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.5
pyteomics==4.7.5
pytest==8.4.2
pytest-cov==7.0.0
pytest-flake8==1.3.0
pytest-pydocstyle==2.4.0
python-dateutil==2.9.0.post0
python-logging-loki==0.3.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.32.5
rfc3339==6.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
setuptools-scm==9.2.0
six==1.17.0
snowballstemmer==3.0.1
SQLAlchemy==2.0.43
stack-data==0.1.4
tabulate==0.9.0
threadpoolctl==3.6.0
tomli==2.3.0
tqdm==4.66.6
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xgboost==1.7.6
xifdr @ file:///C:/Users/fschimweg/workspaces/python/xiFDR-py
xirescore @ file:///C:/Users/fschimweg/workspaces/python/xiRESCORE
xiutilities==1.2.5.post2
yapf==0.31.0
zipp==3.23.0
